push_0102
Added a function that deletes duplicate nodes in the tree. The algorithm does a BFS for duplicate nodes starting from the root's children nodes. The duplicate children are deleted and only the node that has the most immediate children is kept. The algorithm also deletes the note from its parent's children list.